refactor(nuxt add-get-stories): moved it back into the story book folder
